SINGLETON
made singletons inherit from singleton
