+++ LuckTigerIsland/Assets/Scripts/GameMaster/GameMaster.cs
@@ -2,21 +2,7 @@
 using System.Collections.Generic;
 using UnityEngine;
 
-public class GameMaster : MonoBehaviour {
+public class GameMaster : Singleton {
 
-    public static GameMaster instance;
-
-    //To make sure there is only one AudioManager per scene.
-    void Awake()
-    {
-        DontDestroyOnLoad(transform.gameObject);
-        if (instance != null)
-        {
-            Debug.LogError("Only one game master can be present.");
-        }
-        else
-        {
-            instance = this;
-        }
-    }
+    
 }
